+++ src/app/font.ts
@@ -1,3 +1,3 @@
 import { Inter } from 'next/font/google';
 
-export const inter = Inter({ weight: 'variable', subsets: ['cyrillic', 'latin'] });
+export const body = Inter({ weight: 'variable', subsets: ['cyrillic', 'latin'] });