+++ src/features/form/model/types/payload.ts
@@ -1,0 +1,6 @@
+/** Or use a generated graphql type */
+export type RequestPayload = {
+    name: string;
+    email: string;
+    phone: string;
+};