+++ src/pages/ErrorPage/model/type/variant.ts
@@ -1,0 +1,1 @@
+export type Variant = '404' | '500' | 'under-construction';