+++ src/shared/ui/inputs/Button/utils.ts
@@ -1,8 +1,7 @@
 import type { LoaderProps } from 'shared/ui/display/Loader';
-
-export type ButtonVariant = 'primary' | 'secondary';
+import type { ButtonVariant } from './index';
 
 export const MAP_BUTTON_VARIANT_TO_LOADER_VARIANT: Record<ButtonVariant, LoaderProps['variant']> = {
-    primary: 'white',
-    secondary: 'primary',
+    Primary: 'White',
+    Secondary: 'Primary',
 };